requirements.txt
requests==2.32.4
discord.py==2.5.2
scapi==2.1.1
python-dotenv==1.1.1
watchfiles==1.1.0
